Split chain reaction and theme modules
Refactors `ChainReactionChart` by extracting runtime logic, overlay rendering, and public types into dedicated `chainReactionRuntime`, `chainReactionOverlay`, and `chainReactionTypes` modules. Physics server-rendering tests were moved out of the main physics test file into a new `PhysicsCharts.server.test.tsx` to keep tests scoped and reduce oversized files.

Also splits theme serialization (`themeToCSS`/`themeToTokens`) into a standalone `themeSerialization` module and makes `semiotic-themes` a React-free re-export of `semiotic-themes-core` for cleaner package surfaces. Includes small cleanup in package-surface manifest generation and updates file-size policy entries to reflect the new module structure.
